Cleaned up header includes in util

+++ src/util/mpn.cpp
@@ -23,7 +23,7 @@ Revision History:
 
 #define max(a,b)    (((a) > (b)) ? (a) : (b))
 
-typedef uint64 mpn_double_digit;
+typedef unsigned long long mpn_double_digit;
 COMPILE_TIME_ASSERT(sizeof(mpn_double_digit) == 2 * sizeof(mpn_digit));
 
 const mpn_digit mpn_manager::zero = 0;
